fix: Do not mark nodes as physically implausible because of negative disk/spheroid masses
Doing so causes calculation of sizes to be suspended, resulting in sizes being frozen at the value that they had when all components last had positive masses. Removing this allows sizes of components with positive masses to continue to be updated. Those with negative masses are simply ignored during size determination, so have no effect on the results.