fix: Avoid creating duplicate tables for the `massDistributionFiniteResolutionNFW` class
This class was using the standard `hashedDescriptor()` method to create file name suffixes for tabulated solutions. However, that made the suffix depend on the specific density and radius scales of the mass distribution, when the tables are, in fact, computed for the scale-free case. This often resulted in huge numbers of tables being constructed as a new `massDistributionFiniteResolutionNFW` is constructed for each evolution step of each halo.

This fix changes the behavior to use a fixed suffix that depends only on the source code digest.