fix: Make evaluation of UniverseMachine fits more robust
The order of evaluations is changed to avoid floating overflows.

This revision also introduces parameters which control the maximum redshift and minimum halo mass for which UniverseMachine fits will be applied. These default to sensible values for which the UniverseMachine was calibrated, and so avoids applying the model in regions where the fits may be unreliable.