feat: Force recomputation of likelihood for halo mass function if correlations are included
Correlations require evaluation of the cumulative distribution function of very high dimesional multivariate normals. This typically must be done using Monte Carlo integration which can have poor convergence properties in high dimensions. If a chain happens to get stuck in a state that was judged to have very high likelihood because of such Monte Carlo fluctuations, this recomputation can allow it to break out.