fix: Avoid infinite loop when using evaluated parameters that are at a depth of 3 or more in the XML
Previously if an evaluated parameter was at a depth of 3 or greater in the parameter file XML and referenced a parameter with an absolute path, an infinite loop would occur when trying to reach the ultimate parent parameter. This happened because the intermediate subparameter structures were destroyed along the way, making it impossible to traverse back up the tree. We now keep a pointer to the original parent parameters so that we can simply jump to these as needed.