fix: Add an `energy` method to the `massDistributionSphericalDecorator` class
Previously, this class had methods only for the `energyPotential` and `energyKinetic` classes. These, correctly, checked whether the undecorated solution could be used, falling through to the decorated class in such cases. However, since no `energy` method was supplied, if the energy was requested, no check was ever performed to see if the undecorated solution could be used. Instead, therefore, the default approach of a numerical solution would be used, resulting in expensive integrations to evaluate potential and kinetic energies. By adding this `energy` method we can now check if the undecorated solution can be used. In the case of a decorated `massDistributionSphericalTabulated` object (for example), this is much faster as the total energy is tabulated (and so can be returned rapidly), whereas the individual potential and kinetic energies are not (which therefore previously required a slow, numerical evaluation).