feat: Improve thread/process coordination in the creation of CAMB transfer function files
Previously, threads/processes would check for the existance of the transfer function file, and, if it was not found, or was of insufficient extent, would then get an exclusive lock and (re)make te file. Now, after determining that the table is insufficient, we get an exclusive lock, but then test again if the table is sufficient, and only rebuild if it is found to be insuffient on this second check. In this way, if another thread/process had made the table (sufficiently) while our first thread/process was waiting for the exclusive lock, we don't attempt to make the table again. This is important for large parallel jobs where previously each process would create the exact same file in turn, which could lead to very long start up delays.