fix: Correct calculation of merger tree evolution timesteps when using the `starFormationHistoryFixedAges` class
Adds a new method, `timeNext()`, to the `starFormationHistory` class which computes the time of the next star formation history time bin boundary for a given node. A default implementation is provided which simply assumes that the times stored in the star formation history objects are actual times. For the `starFormationHistoryFixedAges` class its own implementation is provided which accounts for the fact that the "times" stored in histories by this class are actually negative ages, and must be corrected to actual times, _and_ compared across all crossing times.