fix: Avoid persistence of temporary parameter files on error
Previously, temporary parameter data was written to `/dev/shm/glcTmp*` files. These files were removed once the main file was opened and ready to receive parameters. If an error occurred during parameter reading the temporary file would be left in place, leading to an accumulation of these files.

This revision calls `unlink()` immediately after opening the temporary file. This removes the link to the file in the file system, but keeps it usable. Once the file is closed, the kernel then removes the data associated with the file. If the code crashes, the kernel cleans up any open file handles, also resulting in the data associated with the file being removed. This avoids accumulation of these files.