fix: Workaround a `gfortran` internal compiler error
Seems to affect functions that make use of the `rootFinder` class. If we mark these as `save`d and `threadprivate` then this avoids the error (probably because no automatic constructor call is generated since these objects are no longer allocated on the stack?). This also has the advantage of being a minor optimization (since we then do not need to construct these `rootFinder`s on every call to the relevant function).