fix: Avoid missing function error if the `FFTLog` tool can not be downloaded
If the `FFTLog` tool can not be download we inject dummy functions for the top-level functions of the tool, such that the code will still compile, but any attempt to call the tool will issue a warning. One top-level function was missing - we now inject a dummy function for that one also.