fix: Restore the previous behavior for concentrations in the reference model
Commit 0917f7ca17 change the settings of the Diemer & Joyce (2019) concentration model used in reference models to have:
```
	  <truncateConcentration value="true" />
	  <includeUpturn         value="false"/>
```
Previously these had not been specified, and so took on the default values - which were the opposites of the above.

This revision reverts this, and now explicitly sets these to:
```
	  <truncateConcentration value="false"/>
	  <includeUpturn         value="true" />
```
in the reference model parameter files.

They _are_ inverted for models with suppressed power spectra as the upturn is problematic in such case.