fix: Correct the calculation of the subresolution angular momentum variance
In the `nodeOperatorHaloAngularMomentumVitvitska2002` class, variance in the angular momentum from subresolution accretion is computed using (for each component of the angular momentum vector):
$$
    J_\mathrm{i}(t_2) = J_\mathrm{i}(t_1) + \left[ \sigma^2 \left\{ J_\mathrm{v}^2(t_2) - J_\mathrm{v}^2(t_1) \right\} f_\mathrm{u} \right]^{1/2} N(0,1),
$$
where $J_\mathrm{v}(t) = M_\mathrm{v}(t) V_\mathrm{v}(t) R_\mathrm{v}(t)$ is the characteristic virial angular momentum, $M_\mathrm{v}(t)$, $V_\mathrm{v}(t)$, and $R_\mathrm{v}(t)$ are the virial mass, velocity, and radius respectively, $M_\mathrm{u}$ is the unresolved mass between times $t_1$ and $t_2$ respectively, $\sigma^2$ represents the variance in angular momentum per unit increase in $J_\mathrm{v}^2$, and $N(0,1)$ is a random variable distributed as a standard normal.

Previously the factor $f_\mathrm{u}$ was computed using:
$$
    f_\mathrm{u} = \left\{ \frac{M(t_1)+M_\mathrm{r}}{M(t_1)} \right\}^2 = \left\{ \frac{M(t_2)-M_\mathrm{u}}{M(t_1)} \right\}^2.
$$
In this, flawed, approach the variance arising from unresolved mass is, incorrectly, not proportional to the unresolved mass accretion (and also has the incorrect scaling with mass).

This patch changes this to:
$$
    f_\mathrm{u} = \frac{M_\mathrm{u}}{M(t_2)},
$$
in which case the variance is proportional to the mass in unresolved accretion (i.e. the ``time'' variable in the Wiener process is just mass).

The old behavior remains an option if `[useOriginalSubresolutionMethod]=true` (which is the default).

Reference models are updated to use the corrected approach.