feat: Allow output of observed positions in lightcones
Adds an option `includeObservedPosition` to the `nodePropertyExtractorLightcone` class, which causes additional output of the "observed" $(x,y,z)$ position - i.e. that which would be inferred from the observed redshift (including the effects of line of sight velocity) and angular position. Specifically:
$$
\mathbf{x}_\mathrm{observed} = \mathbf{x}_\mathrm{true} \frac{D_\chi(z_\mathrm{observed})}{D_\chi(z_\mathrm{true})},
$$
where $\mathbf{x}_\mathrm{true}$ is the true comoving position of the galaxy, $D_\chi(z)$ is the comoving distance to redshift $z$, $z_\mathrm{true}$ is the true cosmological redshift of the galaxy, and $z_\mathrm{observed}$ is the observed redshift including line of sight velocity effects.