fix: Use `floor` instead of `int`
When determining the bin of a histogram into which to accumulate a value we must use `floor()` instead of `int()`. Otherwise, values that should be in bin `-1` (i.e. falling below the first bin of the histogram and so should not be counted) will erroneously be placed into bin `0` - resulting in an excess in that bin.