fix: Avoid double-weighting by stellar/black hole masses
In the stellar/black hoe mass vs. halo mass analysis classes the weighting applied to each galaxy was incorrect. Weighting should be by the log₁₀ of the stellar/black hole mass. This is automatically applied within the `outputAnalysisMeanFunction1D` class. However, the same weighting was being applied directly in the weight operator passed to this classes. This resulted in a double weighting by the log₁₀ of mass, meaning that the quantity computed was actually:
$$
\frac{\langle (\log_{10} M)^2 \rangle}{\langle \log_{10} M \rangle}.
$$
This patch removes the extra weighting from the weight operator such that the correct mean is computed.