fix: Initialize the correct limit
In the `distributionFunction1DNormal` class, the lower limit (if not explicitly specified) was not being initialized - the upper limit was erroneously initialized instead of the lower, this incorrect upper limit initialization would be overridden later, but this left the lower limit in an undefined state. Previously this did not matter as, if the lower limit was not explicitly specified, the value was never used. Now, the lower limit is used as part of the validation checks in the constructor. If this variable happens to contain some non-valid value (since we do not initialize it, its value is undefined), an error can result. This patch now correctly intiializes the lower limit in such cases.