fix: Avoid overwriting earlier stages during constrained tree build
Previously the `mergerTreeWalkerTreeConstruction` class would always begin with the base node of a tree, even if the tree had been already partially built. When doing multi-stage, constrained tree builds, this would cause later stages to simply overwrite the earlier stages (and, therefore, resulted in trees that do not match the imposed constraints).

With this fix, the `mergerTreeWalkerTreeConstruction` class now only starts with the base node if it has no child nodes, otherwise following its usual walking rules.

Additionally, this fix ensures that the `mergerTreeBuilderCole2000` class converts times of _all_ nodes (not just the base node) to critical overdensities before commencing to build. This ensures that if it is handed a partially-built tree, times of all nodes are converted to the critical overdensities used as the internal time variable.