fix: For named parameters in `objectBuilder` directives, add them to the list of recognized parameters for each executable
Drops the requirement that such named parameters be read from the "global parameters" in order to be included in the list of required parameters. The concept of "global parameters" has long-since been retired, and parameters in `objectBuilder` directives can always be read from the top level of the parameter file.