fix: Unlink node from tree before destroying its branch
When ignoring tree branch well-ordering in the `mergerTreeBuilderCole2000` class, if a branch is not well-ordered it is simply pruned away. Previously, the node at the branch tip was not unlinked from the tree prior to pruning, leaving danling pointers to the node (which subsequently led to other problems). The branh tip node is now unlinked before the branch is destroyed.