fix: Fix input parameter validation when multiple parent classes are used
When an object's `Parameters` constructor calls that of its parent class, input parameter validation is disable in the parent class (otherwise this would flag parameters of the child class as not recognized). Previously, if multiple levels of parent class were called in this way, this disabling of validation was broken - it used a simple boolean variable which would be set back to `false` on returning from the grandparent class, which would then allow validation to proceed in the parent class (resulting in false positive warnings about unknown parameters). The boolean is replaced with a counter so that arbitrary levels of parent class can be called, with validation occuring only in the original child class.