fix: Use the correct expansion factor when converting from comoving to physical coordinates
When computing position interpolations for halos, in cases where a subhalo makes a branch jump we are forced to associate a host halo with it that is actually at the subsequent snapshot. When converting the internal position of that host (which is in physical coordinates) to comoving coordinates we must use the expansion factor of that host, even though it will be at a different time from the actual time of the node. Previously we did not do this, which lead to incorrect comoving coordinates which, in some cases, could lead to huge (of order light speed) velocities for nodes, which then lead to problems in solving for the lightcone crossing time.