fix: Make names of soliton properties more distinct
Prepend `soliton` to the names (both for internal meta-properties and output properties) to make is obvious to what these refer, and to avoid name clashes with other parts of the model (e.g. previously `massCore` was used, which is not obviously related to the soliton, and could easily to used by some other component of the model).