fix: Correctly handle deep nested hierarchies of subhalos
When looking for branch jumps, first follow host nodes to the ultimate host, and then move to that ultimate host's descendant, and follow hosts of that until its ultimate host is reached. Previously we simply moved to the immediate host's descendant's host, and followed hosts until the ultimate host was reached. But, if the immediate host was itself a subhalo this could lead to the wrong descendant host being identified (since the `mergerTreeConstructorRead` class currently flattens all deep hierarchies we must always first seek the current ultimate host before descending).